+++ src/components/Dao/Details.tsx
@@ -1,5 +1,5 @@
-import { useDAOData } from "../../../contexts/daoData";
-import H1 from "../../ui/H1";
+import { useDAOData } from "../../contexts/daoData";
+import H1 from "../ui/H1";
 
 function Details() {
   const [{ name, accessControlAddress, moduleAddresses, tokenData }] = useDAOData();
