+++ src/components/Dao/Summary.tsx
@@ -1,7 +1,7 @@
 import { Link } from "react-router-dom";
-import ProposalsList from "./Proposals/ProposalsList";
-import H1 from "../../ui/H1";
-import { SecondaryButton, TextButton } from '../../ui/forms/Button';
+import ProposalsList from "../DAOs/DAO/Proposals/ProposalsList";
+import H1 from "../ui/H1";
+import { SecondaryButton, TextButton } from '../ui/forms/Button';
 
 function Summary() {
   return (
